ecomm-app/requirements.txt
streamlit==1.32.0
databricks-sdk==0.20.0
pandas==2.1.0